Make sure types are same size on 64-bit machines.
git-svn-id: http://svn.digium.com/svn/dahdi/linux/trunk@4742 c8ef98d6-144f-0410-837d-fce3034373d0

+++ drivers/dahdi/dahdi_transcode.c
@@ -231,7 +231,7 @@ __find_free_channel(struct list_head *list, const struct dahdi_transcoder_format
 			}
 		}
 	}
-	return (void*)((match) ? -EBUSY : -ENODEV);
+	return (void*)((long)((match) ? -EBUSY : -ENODEV));
 }
 
 static long dahdi_tc_allocate(struct file *file, unsigned long data)
